refactor: Move locator derivation out of codegen into utils
The locator derivation (toLocatorOptions, getElementLocator, the get*Locator
helpers, isLocatorEqual) maps a recorded element selector to a Playwright-style
locator. It lived under src/codegen/browser, yet the recorder event list and the
browser test editor already depend on it and so reached into the codegen module.
Move it to src/utils/locator so those consumers no longer import from codegen.
No behavior change.
